Refactor client code with internal one.
