--- client/mtls.go
@@ -1,126 +1,0 @@
-package client
-
-import (
-	"context"
-	"crypto/tls"
-	"crypto/x509"
-	"fmt"
-	"log"
-	"os"
-	"time"
-
-	"github.com/reeflective/team/internal/proto"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials"
-)
-
-const (
-	kb = 1024
-	mb = kb * 1024
-	gb = mb * 1024
-
-	// ClientMaxReceiveMessageSize - Max gRPC message size ~2Gb
-	ClientMaxReceiveMessageSize = (2 * gb) - 1 // 2Gb - 1 byte
-
-	defaultTimeout = time.Duration(10 * time.Second)
-)
-
-// connect establishes a working gRPC client connection to the server specified in the configuration.
-func (c *Client) connect(config *Config) (*grpc.ClientConn, error) {
-	tlsConfig, err := getTLSConfig(config.CACertificate, config.Certificate, config.PrivateKey)
-	if err != nil {
-		return nil, err
-	}
-	transportCreds := credentials.NewTLS(tlsConfig)
-	callCreds := credentials.PerRPCCredentials(tokenAuth{token: config.Token})
-	options := []grpc.DialOption{
-		grpc.WithTransportCredentials(transportCreds),
-		grpc.WithPerRPCCredentials(callCreds),
-		grpc.WithBlock(),
-		grpc.WithDefaultCallOptions(grpc.MaxCallRecvMsgSize(ClientMaxReceiveMessageSize)),
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), defaultTimeout)
-	defer cancel()
-	connection, err := grpc.DialContext(ctx, fmt.Sprintf("%s:%d", config.Host, config.Port), options...)
-	if err != nil {
-		return nil, err
-	}
-
-	// Register the core RPC methods
-	c.conn = connection
-	c.rpc = proto.NewTeamClient(c.conn)
-
-	return connection, nil
-}
-
-func getTLSConfig(caCertificate string, certificate string, privateKey string) (*tls.Config, error) {
-	certPEM, err := tls.X509KeyPair([]byte(certificate), []byte(privateKey))
-	if err != nil {
-		log.Printf("Cannot parse client certificate: %v", err)
-		return nil, err
-	}
-
-	// Load CA cert
-	caCertPool := x509.NewCertPool()
-	caCertPool.AppendCertsFromPEM([]byte(caCertificate))
-
-	// Setup config with custom certificate validation routine
-	tlsConfig := &tls.Config{
-		Certificates:       []tls.Certificate{certPEM},
-		RootCAs:            caCertPool,
-		InsecureSkipVerify: true, // Don't worry I sorta know what I'm doing
-		VerifyPeerCertificate: func(rawCerts [][]byte, _ [][]*x509.Certificate) error {
-			return rootOnlyVerifyCertificate(caCertificate, rawCerts)
-		},
-	}
-	return tlsConfig, nil
-}
-
-// rootOnlyVerifyCertificate - Go doesn't provide a method for only skipping hostname validation so
-// we have to disable all of the certificate validation and re-implement everything.
-// https://github.com/golang/go/issues/21971
-func rootOnlyVerifyCertificate(caCertificate string, rawCerts [][]byte) error {
-	roots := x509.NewCertPool()
-	ok := roots.AppendCertsFromPEM([]byte(caCertificate))
-	if !ok {
-		log.Printf("Failed to parse root certificate")
-		os.Exit(3)
-	}
-
-	cert, err := x509.ParseCertificate(rawCerts[0]) // We should only get one cert
-	if err != nil {
-		log.Printf("Failed to parse certificate: " + err.Error())
-		return err
-	}
-
-	// Basically we only care if the certificate was signed by our authority
-	// Go selects sensible defaults for time and EKU, basically we're only
-	// skipping the hostname check, I think?
-	options := x509.VerifyOptions{
-		Roots: roots,
-	}
-	if options.Roots == nil {
-		panic("no root certificate")
-	}
-	if _, err := cert.Verify(options); err != nil {
-		log.Printf("Failed to verify certificate: " + err.Error())
-		return err
-	}
-
-	return nil
-}
-
-type tokenAuth struct {
-	token string
-}
-
-// Return value is mapped to request headers.
-func (t tokenAuth) GetRequestMetadata(ctx context.Context, in ...string) (map[string]string, error) {
-	return map[string]string{
-		"Authorization": "Bearer " + t.token,
-	}, nil
-}
-
-func (tokenAuth) RequireTransportSecurity() bool {
-	return true
-}
